adding the controllers in the autoconfig
